new mandatory argument added

+++ scripts/scudum/util/install.dev.sh
@@ -6,6 +6,11 @@ SLEEP_TIME=3
 
 DIR=$(dirname $(readlink -f $0))
 
+if [ "$DEV_NAME" == "/dev/null" ]; then
+    echo "DEV_NAME not specified, it's required"
+    exit 1
+fi
+
 dd if=/dev/zero of=$DEV_NAME count=1
 
 (echo n; echo p; echo 1; echo ; echo $BOOT_SIZE; echo a; echo 1; echo w) | fdisk -H 255 -S 63 $DEV_NAME
