+++ scripts/scudum/util/make.img.sh
@@ -7,6 +7,11 @@ REBUILD=${REBUILD-0}
 
 DIR=$(dirname $(readlink -f $0))
 
+if [ "$DEV_NAME" == "/dev/null" ]; then
+    echo "DEV_NAME not specified, it's required"
+    exit 1
+fi
+
 if [ "$REBUILD" == "1" ]; then
     dd if=/dev/zero of=$DEV_NAME count=1
