+++ scripts/scudum/util/make.pen.sh
@@ -22,6 +22,11 @@ DEV_ROOT="$DEV_NAME"3
 CUR=$(pwd)
 DIR=$(dirname $(readlink -f $0))
 
+if [ "$DEV_NAME" == "/dev/null" ]; then
+    echo "DEV_NAME not specified, it's required"
+    exit 1
+fi
+
 if [ "$REBUILD" == "1" ]; then
     dd if=/dev/zero of=$DEV_NAME count=1
